requirements.txt
Flask==3.0.0
PyYAML==6.0.1
Werkzeug==3.0.0
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
python-dotenv==1.0.0
requests==2.32.3
